+++ src/Components/CreateImageWizard/validators.ts
@@ -1,4 +1,4 @@
-import type { Partition } from './steps/FileSystem/FileSystemConfiguration';
+import type { Partition } from './steps/FileSystem/FileSystemTable';
 
 export const isAwsAccountIdValid = (awsAccountId: string | undefined) => {
   return (
